be/requirements.txt
croniter==0.3.30
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
python-dateutil==2.8.0
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
